Refactor: Move token generation to `middleware/token` package
