+++ api/internal/server/middleware/token/generate.go
@@ -1,6 +1,7 @@
-package common
+package token
 
 import (
+	cmn "Codex-Backend/api/common"
 	"Codex-Backend/api/internal/domain"
 	"crypto/rand"
 	"encoding/hex"
@@ -12,13 +13,13 @@ import (
 
 func GenerateTokenPair(ID, email string, config domain.TokenConfig) (*domain.TokenPair, error) {
 	if ID == "" {
-		return nil, &Error{Err: errors.New("user ID cannot be empty")}
+		return nil, &cmn.Error{Err: errors.New("user ID cannot be empty")}
 	}
 	if email == "" {
-		return nil, &Error{Err: errors.New("email cannot be empty")}
+		return nil, &cmn.Error{Err: errors.New("email cannot be empty")}
 	}
 	if config.SigningKey == "" {
-		return nil, &Error{Err: errors.New("signing key not configured")}
+		return nil, &cmn.Error{Err: errors.New("signing key not configured")}
 	}
 
 	// Generate access token
@@ -54,7 +55,7 @@ func generateAccessToken(ID, email string, config domain.TokenConfig) (string, t
 	// Generate unique token ID for revocation capability
 	jti, err := generateJTI()
 	if err != nil {
-		return "", time.Time{}, &Error{Err: errors.New("failed to generate token ID: " + err.Error())}
+		return "", time.Time{}, &cmn.Error{Err: errors.New("failed to generate token ID: " + err.Error())}
 	}
 
 	claims := &domain.Claims{
@@ -74,7 +75,7 @@ func generateAccessToken(ID, email string, config domain.TokenConfig) (string, t
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 	tokenString, err := token.SignedString([]byte(config.SigningKey))
 	if err != nil {
-		return "", time.Time{}, &Error{Err: errors.New("token signing failed: " + err.Error())}
+		return "", time.Time{}, &cmn.Error{Err: errors.New("token signing failed: " + err.Error())}
 	}
 
 	return tokenString, expirationTime, nil
@@ -86,7 +87,7 @@ func generateRefreshToken(userID string, config domain.TokenConfig) (string, err
 
 	jti, err := generateJTI()
 	if err != nil {
-		return "", &Error{Err: errors.New("failed to generate refresh token ID: " + err.Error())}
+		return "", &cmn.Error{Err: errors.New("failed to generate refresh token ID: " + err.Error())}
 	}
 
 	// Refresh tokens have minimal claims
@@ -103,7 +104,7 @@ func generateRefreshToken(userID string, config domain.TokenConfig) (string, err
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 	tokenString, err := token.SignedString([]byte(config.SigningKey))
 	if err != nil {
-		return "", &Error{Err: errors.New("refresh token signing failed: " + err.Error())}
+		return "", &cmn.Error{Err: errors.New("refresh token signing failed: " + err.Error())}
 	}
 
 	return tokenString, nil
@@ -120,10 +121,10 @@ func generateJTI() (string, error) {
 
 func DefaultTokenConfig() domain.TokenConfig {
 	return domain.TokenConfig{
-		SigningKey: GetEnvVariable("JWT_SIGN_KEY"),
+		SigningKey: cmn.GetEnvVariable("JWT_SIGN_KEY"),
 		AccessTTL:  30 * time.Minute,
 		RefreshTTL: 7 * 24 * time.Hour,
-		Issuer:     GetEnvVariable("JWT_ISSUER"),
-		Audience:   GetEnvVariable("JWT_AUDIENCE"),
+		Issuer:     cmn.GetEnvVariable("JWT_ISSUER"),
+		Audience:   cmn.GetEnvVariable("JWT_AUDIENCE"),
 	}
 }
